third_party/WebKit/LayoutTests/http/tests/inspector/load-resource-synchronously-encoding.html

<html>
<head>
<script src="inspector-test.js"></script>
<script>
function test()
{
    var phpURL = WebInspector.ParsedURL.completeURL(WebInspector.inspectedPageURL, "resources/load-resource-synchronously-encoding.php");

    function executeTest(charset)
    {
        var data = InspectorFrontendHost.loadResourceSynchronously(phpURL + "?charset=" + charset);
        InspectorTest.addResult("'" + data + "' (" + data.length + ")");
        data = InspectorFrontendHost.loadResourceSynchronously(phpURL + "?charset=" + charset + "&needEncoding=true");
        InspectorTest.addResult("'" + data + "' (" + data.length + ") (with charset)");
    }
    executeTest("US-ASCII");
    executeTest("1251");
    executeTest("1252");
    executeTest("utf-8");
    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests that InspectorFrontendHost.loadResourceSynchronously() correctly loads non-ASCII resources.</p>
</body>
</html>
