third_party/WebKit/LayoutTests/http/tests/inspector/load-resource-synchronously-encoding-expected.txt

Tests that InspectorFrontendHost.loadResourceSynchronously() correctly loads non-ASCII resources.

'US-ASCII: Test' (14)
'US-ASCII: Test' (14) (with charset)
'windows-1251: Тест' (18)
'windows-1251: Тест' (18) (with charset)
'windows-1252: Test Ääšß' (23)
'windows-1252: Test Ääšß' (23) (with charset)
'utf-8: Тест' (11)
'utf-8: Тест' (11) (with charset)
